+++ p2p/protocol/identify/metrics.go
@@ -35,17 +35,73 @@ var (
 		},
 		[]string{"dir"},
 	)
+	connPushSupportTotal = prometheus.NewCounterVec(
+		prometheus.CounterOpts{
+			Namespace: metricNamespace,
+			Name:      "conn_push_support_total",
+			Help:      "Identify Connection Push Support",
+		},
+		[]string{"support"},
+	)
+	protocolsCount = prometheus.NewGauge(
+		prometheus.GaugeOpts{
+			Namespace: metricNamespace,
+			Name:      "protocols_count",
+			Help:      "Protocols Count",
+		},
+	)
+	addrsCount = prometheus.NewGauge(
+		prometheus.GaugeOpts{
+			Namespace: metricNamespace,
+			Name:      "addrs_count",
+			Help:      "Address Count",
+		},
+	)
+	numProtocolsReceived = prometheus.NewHistogram(
+		prometheus.HistogramOpts{
+			Namespace: metricNamespace,
+			Name:      "protocols_received",
+			Help:      "Number of Protocols received",
+			Buckets:   buckets,
+		},
+	)
+	numAddrsReceived = prometheus.NewHistogram(
+		prometheus.HistogramOpts{
+			Namespace: metricNamespace,
+			Name:      "addrs_received",
+			Help:      "Number of addrs received",
+			Buckets:   buckets,
+		},
+	)
 	collectors = []prometheus.Collector{
 		pushesTriggered,
 		identify,
 		identifyPush,
+		connPushSupportTotal,
+		protocolsCount,
+		addrsCount,
+		numProtocolsReceived,
+		numAddrsReceived,
 	}
+	// 1 to 20 and then up to 100 in steps of 5
+	buckets = append(
+		prometheus.LinearBuckets(1, 1, 20),
+		prometheus.LinearBuckets(25, 5, 16)...,
+	)
 )
 
 type MetricsTracer interface {
+	// TriggeredPushes counts IdentifyPushes triggered by event
 	TriggeredPushes(event any)
-	Identify(network.Direction)
-	IdentifyPush(network.Direction)
+
+	// ConnPushSupport counts peers by Push Support
+	ConnPushSupport(identifyPushSupport)
+
+	// IdentifyReceived tracks metrics on receiving an identify response
+	IdentifyReceived(isPush bool, numProtocols int, numAddrs int)
+
+	// IdentifySent tracks metrics on sending an identify response
+	IdentifySent(isPush bool, numProtocols int, numAddrs int)
 }
 
 type metricsTracer struct{}
@@ -90,18 +146,61 @@ func (t *metricsTracer) TriggeredPushes(ev any) {
 	pushesTriggered.WithLabelValues(*tags...).Inc()
 }
 
-func (t *metricsTracer) Identify(dir network.Direction) {
+func (t *metricsTracer) IncrementPushSupport(s identifyPushSupport) {
 	tags := metricshelper.GetStringSlice()
 	defer metricshelper.PutStringSlice(tags)
 
-	*tags = append(*tags, metricshelper.GetDirection(dir))
-	identify.WithLabelValues(*tags...).Inc()
+	*tags = append(*tags, getPushSupport(s))
+	connPushSupportTotal.WithLabelValues(*tags...).Inc()
 }
 
-func (t *metricsTracer) IdentifyPush(dir network.Direction) {
+func (t *metricsTracer) IdentifySent(isPush bool, numProtocols int, numAddrs int) {
 	tags := metricshelper.GetStringSlice()
 	defer metricshelper.PutStringSlice(tags)
 
-	*tags = append(*tags, metricshelper.GetDirection(dir))
-	identifyPush.WithLabelValues(*tags...).Inc()
+	if isPush {
+		*tags = append(*tags, metricshelper.GetDirection(network.DirOutbound))
+		identifyPush.WithLabelValues(*tags...).Inc()
+	} else {
+		*tags = append(*tags, metricshelper.GetDirection(network.DirInbound))
+		identify.WithLabelValues(*tags...).Inc()
+	}
+
+	protocolsCount.Set(float64(numProtocols))
+	addrsCount.Set(float64(numAddrs))
+}
+
+func (t *metricsTracer) IdentifyReceived(isPush bool, numProtocols int, numAddrs int) {
+	tags := metricshelper.GetStringSlice()
+	defer metricshelper.PutStringSlice(tags)
+
+	if isPush {
+		*tags = append(*tags, metricshelper.GetDirection(network.DirInbound))
+		identifyPush.WithLabelValues(*tags...).Inc()
+	} else {
+		*tags = append(*tags, metricshelper.GetDirection(network.DirOutbound))
+		identify.WithLabelValues(*tags...).Inc()
+	}
+
+	numProtocolsReceived.Observe(float64(numProtocols))
+	numAddrsReceived.Observe(float64(numAddrs))
+}
+
+func (t *metricsTracer) ConnPushSupport(support identifyPushSupport) {
+	tags := metricshelper.GetStringSlice()
+	defer metricshelper.PutStringSlice(tags)
+
+	*tags = append(*tags, getPushSupport(support))
+	connPushSupportTotal.WithLabelValues(*tags...).Inc()
+}
+
+func getPushSupport(s identifyPushSupport) string {
+	switch s {
+	case identifyPushSupported:
+		return "supported"
+	case identifyPushUnsupported:
+		return "not supported"
+	default:
+		return "unknown"
+	}
 }